requirements.txt
pandas
flask
numpy
python-dotenv
mysql-connector-python
pymysql
scikit-learn
-e .